+++ src/canary_flux/__init__.py
@@ -17,6 +17,7 @@ from _canary.plugins.subcommands.run import Run
 
 logger = canary.get_logger(__name__)
 
+
 @canary.hookimpl
 def canary_addcommand(parser: canary.Parser) -> None:
     parser.add_command(CanaryFlux())
